+++ packages/commercetools/api-client/src/types/Api.ts
@@ -33,12 +33,13 @@ export interface ProductSearch extends BaseSearch {
   skus?: string[];
   slug?: string;
   id?: string;
-  filters?: Record<string, Filter>;
+  filters?: Filter[];
 }
 
 export interface Filter {
-  options: FilterOption[];
-  type: string;
+  type: AttributeType;
+  name: string;
+  value: any;
 }
 
 export interface FilterOption {