+++ packages/core/core/src/factories/index.ts
@@ -8,3 +8,4 @@ export * from './useUserBillingFactory';
 export * from './useUserOrdersFactory';
 export * from './apiClientFactory';
 export * from './useReviewFactory';
+export * from './useFacetFactory';