+++ packages/shopify/theme/components/Filters.vue
@@ -1,0 +1,3 @@
+<template>
+ <div>This component should be overriden in integration-specific theme.</div>
+</template>