+++ packages/core/docs/contributing/changelog.md
@@ -15,9 +15,10 @@
 
 ### Breaking changes
 
+* **factories**: created useFacet factory and types, changed helpers in category and product page (#4853)
+
 ### Changes
 
 * **factories**: added useReviews factory ([#4775](https://github.com/DivanteLtd/vue-storefront/issues/4775))
 * **factories**: refactored apiClientFactory (#4777)
 * **types**: removed useLocale types (#4777)
-* **factories**: created useFacet factory and types (#4853)