+++ packages/core/docs/contributing/changelog.md
@@ -5,6 +5,16 @@
 
 ### Breaking changes
 
+- #4616
+
+| Place | Before | 2.1.0 |
+|---|---|---|
+| useCartFactory | appliedCoupon ref exists | appliedCoupon removed |
+| useCartFactory | coupon computed returned in useCart | coupon computed removed |
+| useCartFactory | COUPON type in generic | COUPON type removed from generic |
+| types | - | Added AgnosticCoupon type |
+| types | - | `getCoupon` method added to CartGetters interface |
+
 * **factories**: renamed useReviews factory to useReview ([#4800](https://github.com/DivanteLtd/vue-storefront/issues/4800))
 
 ### Changes