+++ packages/core/docs/contributing/changelog.md
@@ -1,6 +1,7 @@
 # Changelog
-
 ## 2.1.0 (not released)
+
+## 2.0.6
 
 ### Breaking changes
 
@@ -19,3 +20,4 @@
 * **factories**: added useReviews factory ([#4775](https://github.com/DivanteLtd/vue-storefront/issues/4775))
 * **factories**: refactored apiClientFactory (#4777)
 * **types**: removed useLocale types (#4777)
+* **factories**: created useFacet factory and types (#4853)